feat(places_provider.dart): add new PlacesNotifier class and providers
Create a new PlacesNotifier class to manage the state of a list of Place objects. This class includes an addPlace method to add a new Place to the state. Also, add two new providers: addPlaceProvider and placesProvider to expose the PlacesNotifier and the current state of the list of Place objects respectively.